Removing processes from model on exit()
When process calls exit() or exit_group() then it will be removed from
ProcessModel. This patch also includes rename of addProcess to
createProcess and similar.

Also, removal from process model means, that it's moved to a different
collection, a collection of terminated processes, which is kept for future
analysis (not yet implemented).